requirements.txt
# Core Framework
streamlit==1.28.1
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Gmail API & Authentication
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1

# AI/ML Libraries
groq==0.4.1
langchain==0.0.335
langchain-community==0.0.7
openai==1.3.7
tiktoken==0.5.2

# Data Processing
pandas==2.1.3
numpy==1.25.2
python-dateutil==2.8.2
pytz==2023.3

# Email Processing
email-validator==2.1.0
beautifulsoup4==4.12.2
html2text==2020.1.16
python-magic==0.4.27

# Caching & Storage
redis==5.0.1
diskcache==5.6.3
sqlalchemy==2.0.23
sqlite3

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyYAML==6.0.1
toml==0.10.2

# HTTP & API
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Logging & Monitoring
loguru==0.7.2
structlog==23.2.0

# Text Processing
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
regex==2023.10.3

# Utilities
click==8.1.7
tqdm==4.66.1
pathlib2==2.3.7.post1
typing-extensions==4.8.0

# Streamlit Components
streamlit-chat==0.1.1
streamlit-authenticator==0.2.3
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4.post3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Security
cryptography==41.0.7
bcrypt==4.1.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# File Processing
python-multipart==0.0.6
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1

# Async Support
asyncio==3.4.3
aiofiles==23.2.1